requirements.txt
configobj~=5.0.6
pandas~=1.3.0
SQLAlchemy~=1.3.19
gspread~=3.7.0
oauth2client~=4.1.3
google-api-python-client~=1.6.7
df2gspread~=1.0.4
python-decouple~=3.4
Flask~=1.1.2
requests~=2.25.1